pyproject.toml
[project]
name = "vuln-agent"
version = "0.1.0"
description = "Deliberately vulnerable multi-user AI agent for red team training"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.28",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
vuln-agent = "main:main"
vuln-debug = "debug:main"
